fix and refactor: se corrigen errores menores y se reorganizan las carpetas del proyecto [Padilla Santiago]
