+++ src/App.css
@@ -35,6 +35,8 @@
 
 .card {
   padding: 2em;
+  margin: 10px;
+  flex: 1;
 }
 
 .read-the-docs {
