+++ src/assets/components/ListaAlumnos.jsx
@@ -2,12 +2,13 @@ import Card from 'react-bootstrap/Card';
 import Button from 'react-bootstrap/Button';
 import Badge from 'react-bootstrap/Badge';
 import { useNavigate } from "react-router-dom";
+import '../../../src/App.css';
 
 function ListaAlumnos({alumnos}) {
   const navigate = useNavigate();
   const listaAlumnos = alumnos.filter(alumno => alumno.estado).map((alumno) => (
     <div key={alumno.lu} className="col-md-4 d-flex">
-      <Card style={{ margin: "10px", flex: 1 }}>
+      <Card className='card'>
         <Card.Body>
           <Card.Title>{alumno.nombre} {alumno.apellido}</Card.Title>
           <Card.Text>
