+++ src/assets/components/VerDetalles.jsx
